compose: Add feature to notify unsubscribed users.
This feature enables users to receive notifications when they are
mentioned in a stream they are not subscribed to. Created the
Notify Them button, made changes to the compose_banner and implemented
communication to the server. The notify_user_about_mention backend
function notifies the specified user about a mention in the specified
stream. The send_message_to_mentioned_user handles the logic for
sending the notification using the Notification Bot and verifies
the type of stream. Added tests to verify the implementation.

Co-authored-by: Isabela Pereira <isabela.p.de.ornelas@tecnico.ulisboa.pt>